+++ GiftTracker.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.27004.2005
+VisualStudioVersion = 15.0.26730.16
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GiftTracker", "GiftTracker\GiftTracker.csproj", "{93728DF8-4CBE-4108-BA5C-A5329D39E41E}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GiftTrackerClasses", "GiftTrackerClasses\GiftTrackerClasses.csproj", "{84DC2D50-C9B1-4E39-A031-7E352A97FB26}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{93728DF8-4CBE-4108-BA5C-A5329D39E41E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{93728DF8-4CBE-4108-BA5C-A5329D39E41E}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{93728DF8-4CBE-4108-BA5C-A5329D39E41E}.Release|Any CPU.Build.0 = Release|Any CPU
+		{84DC2D50-C9B1-4E39-A031-7E352A97FB26}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{84DC2D50-C9B1-4E39-A031-7E352A97FB26}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{84DC2D50-C9B1-4E39-A031-7E352A97FB26}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{84DC2D50-C9B1-4E39-A031-7E352A97FB26}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
