+++ GiftTracker/AddOrEditGiftWindow.xaml.cs
