+++ GiftTracker/DetailsWindow.xaml
