+++ GiftTracker/DetailsWindow.xaml.cs
