+++ GiftTracker/Properties/AssemblyInfo.cs
