+++ GiftTrackerClasses/Gift.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace GiftTrackerClasses
+{
+    public class Gift
+    {
+        public int Id { get; set; }
+        public string Name { get; set; }
+        public byte[] Image { get; set; }
+        public Person Owner { get; set; }
+        public Occasion Occasion { get; set; }
+    }
+}
